更新: AutoPluginManager 热重载
