+++ src/AutoPluginManager/Config.cs
@@ -49,4 +49,5 @@ public class Config
     [JsonProperty("允许自动更新插件")] public bool AutoUpdate = false;
     [JsonProperty("使用Github源")] public bool UseGithubSource = false;
     [JsonProperty("插件排除列表")] public List<string> UpdateBlackList = new();
+    [JsonProperty("热重载升级插件")] public bool AutoReloadPlugin = true;
 }
