+++ src/AutoPluginManager/README.md
@@ -3,10 +3,13 @@
 - 作者: 少司命，Cai
 - 出处: 本仓库
 - 使用指令自动更新服务器的插件(仅本仓库)
-
+- 提供了热重载插件的可能，但此热重载并不算是真正意义上的热重载，已加载的程序集无法卸载，而是依然存在于应用域
+中，只是不在生效!
 ## 更新日志
 
 ```
+v2.0.1.7
+热重载升级插件
 v2.0.1.4
 添加英文翻译
 v2.0.1.3
@@ -45,7 +48,8 @@ V2.0.0.0
 {
   "允许自动更新插件": false,
   "使用Github源": true,
-  "插件排除列表": []
+  "插件排除列表": [],
+  "热重载升级插件": true
 }
 ```
 ## 反馈
